refactor: Исправит мелкие опечатки в именовании
