+++ src/core/cli-command/cli-command.interface.ts
