+++ src/main.cli.ts
@@ -1,9 +1,9 @@
 #!/usr/bin/env node
 
 import CLIApplication from './app/cli.js';
-import HelpCommand from './core/cli-coomand/help.command.js';
-import VersionCommand from './core/cli-coomand/version.command.js';
-import ImportCommand from './core/cli-coomand/import.command.js';
+import HelpCommand from './core/cli-command/help.command.js';
+import VersionCommand from './core/cli-command/version.command.js';
+import ImportCommand from './core/cli-command/import.command.js';
 
 const myManager = new CLIApplication();
 myManager.registerCommand([
